Added a separate component for DDB + DataGrid
