+++ Angular/src/app/app.component.html
@@ -1,81 +1,12 @@
 <div class="dx-viewport demo-container">
   <div class="row">
     <p>DropDownBox with search and embedded DataGrid</p>
-    <dx-drop-down-box
-      width="40vw"
-      #dropDownBox
-      [dataSource]="dropDownBoxDataSource"
-      [value]="dropDownValue"
-      valueExpr="OrderNumber"
-      [displayExpr]="gridBoxDisplayExpr"
-      [openOnFieldClick]="false"
-      [showClearButton]="true"
-      [acceptCustomValue]="true"
-      [(opened)]="gridBoxOpened"
-      placeholder="Select a value..."
-      valueChangeEvent=""
-      (onInput)="onInput($event)"
-      (onOptionChanged)="onOptionChanged($event)"
-      (onValueChanged)="onDropDownValueChanged($event)"
-      (onKeyDown)="onDropDownBoxKeyDown($event)"
-      (onOpened)="onOpened($event)"
-      (onClosed)="onClosed($event)"
-    >
-      <dxo-drop-down-box-drop-down-options
-        [height]="400"
-      ></dxo-drop-down-box-drop-down-options>
-
-      <div *dxTemplate="let data of 'content'">
-        <dx-data-grid
-          #dataGrid
-          [dataSource]="dataSource"
-          height="100%"
-          width="100%"
-          [focusedRowEnabled]="true"
-          [(focusedRowKey)]="focusedRowKey"
-          [focusedRowIndex]="focusedRowIndex"
-          [(selectedRowKeys)]="selectedRowKeys"
-          [autoNavigateToFocusedRow]="false"
-          [remoteOperations]="true"
-          [columnAutoWidth]="true"
-          (onKeyDown)="dataGridKeyDown($event)"
-          (onContentReady)="dataGridContentReady($event)"
-          (onFocusedRowChanged)="onFocusedRowChanged($event)"
-          (onSelectionChanged)="onSelectionChanged($event)"
-        >
-          <dxi-data-grid-column
-            dataField="OrderNumber"
-            caption="ID"
-            dataType="number"
-          ></dxi-data-grid-column>
-          <dxi-data-grid-column
-            dataField="OrderDate"
-            dataType="date"
-          ></dxi-data-grid-column>
-          <dxi-data-grid-column
-            dataField="StoreState"
-            dataType="string"
-          ></dxi-data-grid-column>
-          <dxi-data-grid-column
-            dataField="StoreCity"
-            dataType="string"
-          ></dxi-data-grid-column>
-          <dxi-data-grid-column
-            dataField="Employee"
-            dataType="string"
-          ></dxi-data-grid-column>
-          <dxi-data-grid-column dataField="SaleAmount" dataType="number">
-            <dxo-format type="currency" [precision]="2"></dxo-format>
-          </dxi-data-grid-column>
-          <dxo-data-grid-paging
-            [enabled]="true"
-            [pageSize]="10"
-          ></dxo-data-grid-paging>
-          <dxo-data-grid-selection mode="single"></dxo-data-grid-selection>
-          <dxo-data-grid-scrolling mode="virtual"></dxo-data-grid-scrolling>
-        </dx-data-grid>
-      </div>
-    </dx-drop-down-box>
+    <app-drop-down-grid
+      [selectedRowKey]="35709"
+      [dataSource]="dataSource"
+      [dropDownBoxDataSource]="dropDownBoxDataSource"
+      [searchTimeout]="searchTimeout"
+    ></app-drop-down-grid>
   </div>
 
   <div class="options">
